+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/BackgroundTaskRunStatus.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util
+{
+    public enum BackgroundTaskRunStatus
+    {
+        NotStarted,
+        Running,
+        Completed,
+        Cancelled,
+        Failed,
+        Unknown
+    }
+}
